feat: add race and reset btn functionality, also add modal text after car's win

+++ async-race/src/scripts/components/carCreator.ts
@@ -1,6 +1,7 @@
 import { ApiQuery } from '../api/apiQuery';
 import { EventEmitter } from '../emitter/emitter';
-import { CarData } from '../types/dataTypes';
+import { Race } from '../race/race';
+import { CarData, EngineData } from '../types/dataTypes';
 import { createSvg } from '../util/createSvg';
 import { Creator } from './creator';
 
@@ -21,6 +22,8 @@ export class CarCreator extends Creator {
   private color = this.carData.color;
   private name = this.carData.name;
   private isShown = false;
+  public isDrive = false;
+  private startPos = 25;
 
   constructor(protected emitter: EventEmitter, private carData: CarData) {
     super(emitter);
@@ -39,7 +42,7 @@ export class CarCreator extends Creator {
     this.btnStop = Creator.renderElem(this.carEngine, 'div', ['car__stop', 'inactive'], 'B');
 
     this.carRoad = Creator.renderElem(this.carWrapper, 'div', ['car__road']);
-    this.carSvg = createSvg(this.color, ['car__instance']);
+    this.carSvg = createSvg(this.color, ['car__instance', 'race']);
     this.carRoad.append(this.carSvg);
     this.carFlag = Creator.renderElem(this.carRoad, 'div', ['car__flag']);
 
@@ -49,16 +52,28 @@ export class CarCreator extends Creator {
   private bindEvents(): void {
     this.btnRemove.addEventListener('click', () => this.deleteCar());
     this.btnSelect.addEventListener('click', () => this.selectCar());
+
+    this.btnStart.addEventListener('click', () => {
+      this.startCar();
+    });
+
+    this.btnStop.addEventListener('click', () => {
+      this.stopCar();
+    });
   }
 
   private deleteCar(): void {
-    this.carWrapper.remove();
-    ApiQuery.delete('garage', this.id);
-    this.emitter.emit('deleteCar', this.id);
+    if (!this.btnRemove.classList.contains('inactive')) {
+      this.carWrapper.remove();
+      ApiQuery.delete('garage', this.id);
+      this.emitter.emit('deleteCar', this.id);
+    }
   }
 
   private selectCar(): void {
-    this.emitter.emit('selectCar', { id: this.id, color: this.color, name: this.name });
+    if (!this.btnSelect.classList.contains('inactive')) {
+      this.emitter.emit('selectCar', { id: this.id, color: this.color, name: this.name });
+    }
   }
 
   public updateCar(data: CarData): void {
@@ -68,6 +83,37 @@ export class CarCreator extends Creator {
     this.carSvg.replaceWith(newSvg);
     this.carSvg = newSvg;
     this.carName.textContent = this.name;
+  }
+
+  public async startCar(): Promise<void> {
+    if (!this.btnStart.classList.contains('inactive')) {
+      this.emitter.emit('disableSelect', this.id);
+      this.btnStart.classList.add('inactive');
+      this.btnSelect.classList.add('inactive');
+      this.btnRemove.classList.add('inactive');
+      const data = await ApiQuery.engine<EngineData>(this.id, 'started');
+      data.id = this.id;
+      data.name = this.name;
+      data.flag = this.carFlag;
+      data.car = this.carSvg;
+      Race.start(data, this.emitter);
+      this.btnStop.classList.remove('inactive');
+      this.isDrive = true;
+    }
+  }
+
+  public async stopCar(): Promise<void> {
+    if (!this.btnStop.classList.contains('inactive')) {
+      this.emitter.emit('enableSelect', this.id);
+      ApiQuery.engine(this.id, 'stopped');
+      Race.stop(this.id);
+      Race.reset(this.carSvg, this.startPos);
+      this.btnStop.classList.add('inactive');
+      this.btnStart.classList.remove('inactive');
+      this.btnSelect.classList.remove('inactive');
+      this.btnRemove.classList.remove('inactive');
+      this.isDrive = false;
+    }
   }
 
   public show(): void {
